nixos/installation-device: don't include clone-config
The configuration used to build an installation medium isn't really
suitable to be (re)used when building a real system.

For example, it enables automatic root login without a password on the
console, installs sshd, but prevents it from starting on boot.

This can be very misleading when users try to derive their real system
from it - they should use nixos-generate-config instead, as described in
the manual.

Closes #63576